+++ cli/cli/commands/enclave/inspect/inspect.go
@@ -21,6 +21,7 @@ import (
 	"sort"
 	"strings"
 	"text/tabwriter"
+	"unicode/utf8"
 )
 
 const (
@@ -33,14 +34,18 @@ const (
 	tabWriterPadchar  = ' '
 	tabWriterFlags    = 0
 
-	guidHeader             = "GUID"
-	nameHeader             = "Name"
-	hostPortBindingsHeader = "HostPortBindings"
+	headerWidthChars = 100
+	headerPadChar = "="
 )
 
 var defaultKurtosisLogLevel = logrus.InfoLevel.String()
 var positionalArgs = []string{
 	enclaveIdArg,
+}
+
+var enclaveObjectPrintingFuncs = map[string]func(ctx context.Context, dockerManager *docker_manager.DockerManager, enclaveId string) error {
+	"Interactive REPLs": printInteractiveRepls,
+	"User Services": printUserServices,
 }
 
 var InspectCmd = &cobra.Command{
@@ -90,64 +95,49 @@ func run(cmd *cobra.Command, args []string) error {
 		dockerClient,
 	)
 
-	labels := getLabelsForListEnclaveUserServices(enclaveId)
+	headersWithPrintErrs := []string{}
+	for header, printingFunc := range enclaveObjectPrintingFuncs {
+		numRunesInHeader := utf8.RuneCountInString(header) + 2	// 2 because there will be a space before and after the header
+		numPadChars := (headerWidthChars - numRunesInHeader) / 2
+		padStr := strings.Repeat(headerPadChar, numPadChars)
+		fmt.Println(fmt.Sprintf("%v %v %v", padStr, header, padStr))
 
-	containers, err := dockerManager.GetContainersByLabels(ctx, labels, true)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting containers by labels: '%+v'", labels)
+		if err := printingFunc(ctx, dockerManager, enclaveId); err != nil {
+			logrus.Error(err)
+			headersWithPrintErrs = append(headersWithPrintErrs, header)
+		}
+		fmt.Println("")
 	}
 
-	if containers != nil {
-		tabWriter := tabwriter.NewWriter(os.Stdout, tabWriterMinwidth, tabWriterTabwidth, tabWriterPadding, tabWriterPadchar, tabWriterFlags)
-		fmt.Fprintln(tabWriter, guidHeader + "\t" + nameHeader + "\t" + hostPortBindingsHeader)
-		sortedContainers, err := getContainersSortedByGUID(containers)
-		if err != nil {
-			return stacktrace.Propagate(err, "An error occurred getting containers sorted by GUID")
-		}
-		for _, container := range sortedContainers {
-			containerGUIDLabel, found := container.GetLabels()[enclave_object_labels.GUIDLabel]
-			if !found {
-				return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", enclave_object_labels.GUIDLabel, container.GetId(), container.GetLabels())
-			}
-			hostPortBindingsStrings := getContainerHostPortBindingStrings(container)
-
-			var firstHostPortBinding string
-			if hostPortBindingsStrings != nil  {
-				firstHostPortBinding = hostPortBindingsStrings[0]
-				hostPortBindingsStrings = hostPortBindingsStrings[1:]
-			}
-			line := containerGUIDLabel + "\t" + container.GetName() + "\t" + firstHostPortBinding
-			fmt.Fprintln(tabWriter, line)
-
-			for _, hostPortBindingsString := range hostPortBindingsStrings {
-				line = "\t\t" + hostPortBindingsString
-				fmt.Fprintln(tabWriter, line)
-			}
-		}
-		tabWriter.Flush()
+	if len(headersWithPrintErrs) > 0 {
+		return stacktrace.NewError(
+			"Errors occurred printing the following enclave elements: %v",
+			strings.Join(headersWithPrintErrs, ", "),
+		)
 	}
 
 	return nil
 }
 
-func getContainerHostPortBindingStrings(container *types.Container) []string {
-	var allHosPortBindings []string
-	hostPortBindings := container.GetHostPortBindings()
-	for hostPortBindingKey, hostPortBinding := range hostPortBindings {
-		hostPortBindingString := fmt.Sprintf("%v -> %v:%v", hostPortBindingKey, hostPortBinding.HostIP, hostPortBinding.HostPort)
-		allHosPortBindings = append(allHosPortBindings, hostPortBindingString)
-	}
-	return allHosPortBindings
-}
-
 // ====================================================================================================
 // 									   Private helper methods
 // ====================================================================================================
-func getLabelsForListEnclaveUserServices(enclaveId string) map[string]string {
-	labels := map[string]string{}
-	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeUserServiceContainer
-	labels[enclave_object_labels.EnclaveIDContainerLabel] = enclaveId
-	return labels
+func getTabWriterForPrinting() *tabwriter.Writer {
+	return tabwriter.NewWriter(
+		os.Stdout,
+		tabWriterMinwidth,
+		tabWriterTabwidth,
+		tabWriterPadding,
+		tabWriterPadchar,
+		tabWriterFlags,
+	)
+}
+
+func writeElemsToTabWriter(writer *tabwriter.Writer, elems... string) {
+	fmt.Fprintln(
+		writer,
+		strings.Join(elems, "\t"),
+	)
 }
 
 func getContainersSortedByGUID(containers []*types.Container) ([]*types.Container, error) {