+++ cli/cli/commands/sandbox/sandbox.go
@@ -130,5 +130,3 @@ func run(cmd *cobra.Command, args []string) error {
 
 	return nil
 }
-
-