+++ cli/cli/enclave_manager/enclave_context/enclave_context.go
@@ -8,6 +8,7 @@ package enclave_context
 import (
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	"github.com/kurtosis-tech/kurtosis-core/commons/object_labels_providers"
 	"github.com/kurtosis-tech/kurtosis-core/commons/object_name_providers"
 	"net"
 )
@@ -25,10 +26,12 @@ type EnclaveContext struct {
 	dockerManager *docker_manager.DockerManager
 
 	objNameProvider *object_name_providers.EnclaveObjectNameProvider
+
+	objLabelProvider *object_labels_providers.EnclaveObjectLabelsProvider
 }
 
-func NewEnclaveContext(enclaveId string, networkId string, networkIpAndMask *net.IPNet, apiContainerId string, apiContainerIpAddr net.IP, apiContainerHostPortBinding *nat.PortBinding, dockerManager *docker_manager.DockerManager, objNameProvider *object_name_providers.EnclaveObjectNameProvider) *EnclaveContext {
-	return &EnclaveContext{enclaveId: enclaveId, networkId: networkId, networkIpAndMask: networkIpAndMask, apiContainerId: apiContainerId, apiContainerIpAddr: apiContainerIpAddr, apiContainerHostPortBinding: apiContainerHostPortBinding, dockerManager: dockerManager, objNameProvider: objNameProvider}
+func NewEnclaveContext(enclaveId string, networkId string, networkIpAndMask *net.IPNet, apiContainerId string, apiContainerIpAddr net.IP, apiContainerHostPortBinding *nat.PortBinding, dockerManager *docker_manager.DockerManager, objNameProvider *object_name_providers.EnclaveObjectNameProvider, objLabelProvider *object_labels_providers.EnclaveObjectLabelsProvider) *EnclaveContext {
+	return &EnclaveContext{enclaveId: enclaveId, networkId: networkId, networkIpAndMask: networkIpAndMask, apiContainerId: apiContainerId, apiContainerIpAddr: apiContainerIpAddr, apiContainerHostPortBinding: apiContainerHostPortBinding, dockerManager: dockerManager, objNameProvider: objNameProvider, objLabelProvider: objLabelProvider}
 }
 
 func (enclaveCtx *EnclaveContext) GetEnclaveID() string {
@@ -61,4 +64,9 @@ func (enclaveCtx *EnclaveContext) GetDockerManager() *docker_manager.DockerManag
 
 func (enclaveCtx *EnclaveContext) GetObjectNameProvider() *object_name_providers.EnclaveObjectNameProvider {
 	return enclaveCtx.objNameProvider
+}
+
+
+func (enclaveCtx *EnclaveContext) GetObjectLabelsProvider() *object_labels_providers.EnclaveObjectLabelsProvider {
+	return enclaveCtx.objLabelProvider
 }