+++ cli/docs/changelog.md
@@ -2,6 +2,7 @@
 ### Features
 * Added `enclave new` command to create a new enclave
 * Upgrade Kurtosis Core Engine Libs to v0.6.0 which adds `Network` type
+* Print interactive REPLs in `enclave inspect`
 
 # 0.3.1
 ### Fixes